Se borra el usuario y se agrega el monto acumulado en el listado de los movimientos
