+++ src/js/caja/components/CajaMain.js
@@ -1,9 +1,12 @@
 import React, { Component } from 'react';
-
+import { connect } from 'react-redux';
+import propTypes from 'prop-types';
 
 import CajaActionBar from './CajaActionBar';
 import SearchCajaModal from './search/SearchCajaModal';
 import ListadoMovimientosTable from './ListadoMovimientosTable';
+
+import { FETCH_MOVIMIENTOS_CAJA } from '../actionTypes';
 
 class CajaMain extends Component {
     constructor(props) {
@@ -15,6 +18,7 @@ class CajaMain extends Component {
 
         this.openSearchCajaModal = this.openSearchCajaModal.bind(this);
         this.closeSearchCajaModal = this.closeSearchCajaModal.bind(this);
+        this.props.fetchMovimientosCaja();
     }
 
     openSearchCajaModal() {
@@ -30,11 +34,14 @@ class CajaMain extends Component {
             <div className='ibox-content'>
                 <div className='pull-right'>
                     <CajaActionBar
+                      montoAcumulado={
+                        this.props.movimientos[0].monto_acumulado
+                    }
                       openSearchCajaModal={ this.openSearchCajaModal }
                     />
                 </div>
                 <div className='clearfix' />
-                <ListadoMovimientosTable />
+                <ListadoMovimientosTable movimientos={ this.props.movimientos } />
                 <SearchCajaModal
                   modalOpened={ this.state.modalOpened }
                   closeModal={ this.closeSearchCajaModal }
@@ -44,4 +51,24 @@ class CajaMain extends Component {
     }
 }
 
-export default CajaMain;
+const { array, func } = propTypes;
+
+CajaMain.propTypes = {
+    movimientos: array.isRequired,
+    fetchMovimientosCaja: func.isRequired,
+};
+
+function mapStateToProps(state) {
+    return {
+        movimientos: state.cajaReducer.movimientos,
+    };
+}
+
+function mapDispatchToProps(dispatch) {
+    return {
+        fetchMovimientosCaja: () =>
+            dispatch({ type: FETCH_MOVIMIENTOS_CAJA }),
+    };
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(CajaMain);
